Refactor: SearchTemplate 위치 이동
