+++ src/pages/SearchPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import SearchTemplate from '~/components/SearchTemplate';
+import SearchTemplate from '~/components/templates/SearchTemplate';
 import useSearchAll from '~/hooks/api/search/useSearchAll';
 import useSearchUsers from '~/hooks/api/search/useSearchUsers';
 import { userState } from '~/recoil/login/atoms';
